yolo-v3/requirements.txt
pillow
torch>=0.4.1
numpy
torchvision
